PublicCalls & Jasmine
PublicCalls is a node I am going to be implementing every grabber that
needs to be public.

e.g.
isAtTree() will be put in PublicCalls because WalkToTrees & ChopTree
will need it to verify, while everything else is private in their own
class.

Jasmine is a valid node that is here to stay, but will go unused.  Made
for my baby.